requirements.txt
streamlit==1.32.2
python-dotenv==1.0.1
langchain-core==0.1.31

# Required for LangChain functionality
langchain==0.1.12
langchain-openai==0.0.8
langchain-community==0.0.28
langchain-qdrant==0.0.1
qdrant-client==1.8.0

# Required for PDF processing
pypdf==4.1.0
unstructured[pdf]==0.12.5
pdf2image==1.17.0
pytesseract==0.3.10
poppler-utils
camelot-py
opencv-python